requirement.txt.txt
streamlit
pyserini
faiss-cpu
sentence-transformers
datasets
numpy